add capability to assign NODE_ENV; include node_modules in gulp dist task; some more gulp fixes; config middleware fixes

+++ source/configMiddleware.js
@@ -14,7 +14,9 @@ switch (mode) {
     case 'json':
         cfgFileExt = '.json';
         break;
-    case 'overwrite': case 'concat': default:
+    case 'overwrite':
+    case 'concat':
+    default:
         cfgFileExt = process.env.CFG_EXT || '.js';
         if (cfgFileExt.substr(0, 1) !== '.') {
             cfgFileExt = '.' + cfgFileExt;
@@ -26,44 +28,44 @@ var cfgDir = path.join(__dirname, '/../public/configs');
 var defaultCfgFile = cfgDir + '/config' + cfgFileExt;
 var randomCfgFile = cfgDir + '/' + require('crypto').randomBytes(24).toString('hex') + cfgFileExt;
 
-if (nodeEnv = process.env.NODE_ENV) {
-    var envCfgFile = cfgDir + '/' + nodeEnv + cfgFileExt;
-    if (fs.statSync(envCfgFile).isFile()) {
-        switch (mode) {
-            case 'json':
-                var defaultCfg = jsonfile.readFileSync(defaultCfgFile);
-                var envCfg = jsonfile.readFileSync(envCfgFile);
-                var config = _.extend(defaultCfg, envCfg);
-                jsonfile.writeFileSync(randomCfgFile, config);
-                break;
-            case 'nodejs':
-                var defaultCfg = require(defaultCfgFile);
-                var envCfg = require(envCfgFile);
-                var config = _.extend(defaultCfg, envCfg);
-                fs.writeFileSync(randomCfgFile, 'var envConfig = ' + JSON.stringify(config) + ';\nmodule.exports = envConfig;');
-                break;
-            case 'concat':
-                var defaultCfg = fs.readFileSync(defaultCfgFile);
-                fs.writeFile(randomCfgFile, defaultCfg, function () {
-                    var envCfg = fs.readFileSync(envCfgFile);
-                    fs.appendFileSync(randomCfgFile, '\n' + envCfg);
-                });
-                break;
-            case 'overwrite':
-            default:
+var nodeEnv = process.env.NODE_ENV;
+var envCfgFile = cfgDir + '/' + nodeEnv + cfgFileExt;
+if (nodeEnv && fs.statSync(envCfgFile).isFile()) {
+    switch (mode) {
+        case 'json':
+            var defaultCfg = jsonfile.readFileSync(defaultCfgFile);
+            var envCfg = jsonfile.readFileSync(envCfgFile);
+            var config = _.extend(defaultCfg, envCfg);
+            jsonfile.writeFileSync(randomCfgFile, config);
+            break;
+        case 'nodejs':
+            var defaultCfg = require(defaultCfgFile);
+            var envCfg = require(envCfgFile);
+            var config = _.extend(defaultCfg, envCfg);
+            fs.writeFileSync(randomCfgFile, 'var envConfig = ' + JSON.stringify(config) + ';\nmodule.exports = envConfig;');
+            break;
+        case 'concat':
+            var defaultCfg = fs.readFileSync(defaultCfgFile);
+            fs.writeFile(randomCfgFile, defaultCfg, function () {
                 var envCfg = fs.readFileSync(envCfgFile);
-                fs.writeFileSync(randomCfgFile, envCfg);
-                break;
-        }
-        ondeath(function () {
-            fs.unlinkSync(randomCfgFile);
-            process.exit();
-        });
+                fs.appendFileSync(randomCfgFile, '\n' + envCfg);
+            });
+            break;
+        case 'overwrite':
+        default:
+            var envCfg = fs.readFileSync(envCfgFile);
+            fs.writeFileSync(randomCfgFile, envCfg);
+            break;
     }
 } else {
-    var config = fs.readFileSync(defaultCfg);
+    var config = fs.readFileSync(defaultCfgFile);
     fs.writeFileSync(randomCfgFile, config);
 }
+
+ondeath(function () {
+    fs.unlinkSync(randomCfgFile);
+    process.exit();
+});
 
 var configMiddleware = function (req, res, next) {
     if (req.originalUrl === "/configs/config.json" || req.originalUrl === "/configs/config.js") {
